requirements.txt
transformers==4.35.2
fastapi==0.104.1
# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cpu
# Добавить команду в readme
python-dotenv==1.0.0
